# lib/api-zod/src/generated/types/createDepenseDetaillee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Ferme Familiale - Gestion de la ferme avicole
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDepenseDetaillee {
  designation: string;
  quantite: number;
  prixUnitaire: number;
}