# lib/api-zod/src/generated/types/createSortieCarburantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Ferme Familiale - Gestion de la ferme avicole
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSortieCarburantBody {
  date: Date;
  montant: number;
}